@@ -29,11 +29,11 @@
     "webpack": "^1.8.5"
   },
   "devDependencies": {
-    "babel-eslint": "^3.0.1",
+    "babel-eslint": "^3.1.2",
     "chai": "^2.3.0",
-    "eslint": "^0.19.0",
-    "eslint-loader": "^0.10.0",
-    "eslint-plugin-react": "2.1.1",
+    "eslint": "^0.21.0",
+    "eslint-loader": "^0.11.2",
+    "eslint-plugin-react": "2.3.0",
     "http-proxy": "^1.10.1",
     "isparta": "^3.0.2",
     "isparta-instrumenter-loader": "^0.2.1",
