@@ -8,31 +8,31 @@ class Navbar extends React.Component {
 
   render() {
     return (
-      <nav className="navbar navbar-inverse navbar-static-top">
-        <div className="container">
-          <div className="navbar-header">
+      <nav className='navbar navbar-inverse navbar-static-top'>
+        <div className='container'>
+          <div className='navbar-header'>
             <button
-              type="button"
-              className="navbar-toggle collapsed"
-              data-toggle="collapse"
-              data-target="#navbar"
-              aria-expanded="false"
-              aria-controls="navbar"
+              type='button'
+              className='navbar-toggle collapsed'
+              data-toggle='collapse'
+              data-target='#navbar'
+              aria-expanded='false'
+              aria-controls='navbar'
             >
-              <span className="sr-only">Toggle navigation</span>
-              <span className="icon-bar"></span>
-              <span className="icon-bar"></span>
-              <span className="icon-bar"></span>
+              <span className='sr-only'>Toggle navigation</span>
+              <span className='icon-bar'></span>
+              <span className='icon-bar'></span>
+              <span className='icon-bar'></span>
             </button>
-            <Link to="search" className="navbar-brand">
-              <span className="coat-of-arms"/>
+            <Link to='search' className='navbar-brand'>
+              <span className='coat-of-arms'/>
               Koulurekisteri
             </Link>
           </div>
-          <div id="navbar" className="navbar-collapse collapse">
-            <ul className="nav navbar-nav">
-                <NavTab to="search">Etsi koulua</NavTab>
-                <NavTab to="about">Tietoa palvelusta</NavTab>
+          <div id='navbar' className='navbar-collapse collapse'>
+            <ul className='nav navbar-nav'>
+              <NavTab to='search'>Etsi koulua</NavTab>
+              <NavTab to='about'>Tietoa palvelusta</NavTab>
             </ul>
           </div>
         </div>
