@@ -1,5 +1,6 @@
 'use strict';
 
 module.exports = require('./make-webpack-config')({
-  devtool: 'cheap-module-eval-source-map'
+  devtool: 'cheap-module-eval-source-map',
+  lint: true
 });
